refactor(ios): migrate budget line detail to push navigation (PUL-209)
Replace bottom-sheet flow with full-page push inside the parent NavigationStack
so users can drill from envelope → transaction edit → save without losing
context. Sheet-then-sheet chains caused visible reopen flicker and dropped the
user back on the budget list after a single tx edit.

- New `BudgetLinePushRoute` enum (Hashable, ID-based) with cases for line
  detail, edit transaction, add allocated transaction. Resolver in
  `BudgetDetailsView` injects the shared `BudgetDetailsViewModel` via
  `.environment(viewModel)` on the destination view; pages re-resolve their
  model reactively from Observation.
- `BudgetLineDetailSheet` → `BudgetLineDetailPage`. Drops internal
  NavigationStack + standardSheetPresentation + SheetCloseButton; auto-pops
  via `dismiss()` when the line is removed externally.
- New `EditTransactionPage` and `AddAllocatedTransactionPage` push pages.
  Form rendered inline via private subviews; pure helpers extracted to
  `EditTransactionLogic` / `AddAllocatedTransactionLogic` enums (testable
  without bootstrapping SwiftUI).
- Free transaction tap also pushes to `EditTransactionPage` (was sheet).
- Drop `EditTransactionSheet` + `AddAllocatedTransactionSheet` (orphans).
  Drop sheet enum cases `lineDetail`, `addAllocatedTransaction`,
  `editTransaction`. Drop `pendingTransactionDeletion` + `handleSheetDismiss`
  (no more dismiss-then-toast pattern; toast fired inline before pop).
- Hero amounts use `asAmount` / `asCurrency` (2 decimals) per spec.
- Row component extracted to `BudgetLineDetailTransactionRow.swift`; helper
  views moved to private extensions to keep the page struct lean.
- Test file moved + migrated to call `EditTransactionLogic.*` instead of
  `EditTransactionSheet.*`.

Note: temporary hardcoded bottom padding on the page's "Ajouter" button to
clear `MainTabView`'s overlay tab bar. Follow-up commit will refactor
`MainTabView` to use `safeAreaInset` for the custom tab bar so all push pages
auto-clear it without per-page padding.
