+++ ios/PulpeTests/Features/Budgets/BudgetDetails/EditTransactionLogicTests.swift
@@ -2,51 +2,14 @@ import Foundation
 @testable import Pulpe
 import Testing
 
-@Suite("EditTransactionSheet Tests")
+@Suite("EditTransactionLogic Tests")
 @MainActor
-struct EditTransactionSheetTests {
-    // MARK: - Dependencies
-
-    @Test
-    func dependencies_updateTransaction_passesCorrectIdAndData() async throws {
-        nonisolated(unsafe) var receivedId: String?
-        nonisolated(unsafe) var receivedData: TransactionUpdate?
-        let expectedTransaction = TestDataFactory.createTransaction(
-            id: "tx-1",
-            name: "Updated",
-            amount: 200,
-            kind: .income
-        )
-
-        let deps = EditTransactionDependencies(
-            updateTransaction: { id, data in
-                receivedId = id
-                receivedData = data
-                return expectedTransaction
-            }
-        )
-
-        let result = try await deps.updateTransaction("tx-1", TransactionUpdate(
-            name: "Updated",
-            amount: 200,
-            kind: .income,
-            transactionDate: TestDataFactory.fixedDate
-        ))
-
-        #expect(receivedId == "tx-1")
-        #expect(receivedData?.name == "Updated")
-        #expect(receivedData?.amount == 200)
-        #expect(receivedData?.kind == .income)
-        #expect(receivedData?.transactionDate == TestDataFactory.fixedDate)
-        #expect(result.id == "tx-1")
-        #expect(result.name == "Updated")
-    }
-
+struct EditTransactionLogicTests {
     // MARK: - Form Validation Logic
 
     @Test
     func canSubmit_validInputs_returnsTrue() {
-        let result = EditTransactionSheet.isFormValid(
+        let result = EditTransactionLogic.isFormValid(
             name: "Groceries",
             amount: 50,
             isLoading: false
@@ -57,7 +20,7 @@ struct EditTransactionSheetTests {
 
     @Test
     func canSubmit_emptyName_returnsFalse() {
-        let result = EditTransactionSheet.isFormValid(
+        let result = EditTransactionLogic.isFormValid(
             name: "",
             amount: 50,
             isLoading: false
@@ -68,7 +31,7 @@ struct EditTransactionSheetTests {
 
     @Test
     func canSubmit_whitespaceOnlyName_returnsFalse() {
-        let result = EditTransactionSheet.isFormValid(
+        let result = EditTransactionLogic.isFormValid(
             name: "   ",
             amount: 50,
             isLoading: false
@@ -79,7 +42,7 @@ struct EditTransactionSheetTests {
 
     @Test
     func canSubmit_zeroAmount_returnsFalse() {
-        let result = EditTransactionSheet.isFormValid(
+        let result = EditTransactionLogic.isFormValid(
             name: "Groceries",
             amount: 0,
             isLoading: false
@@ -90,7 +53,7 @@ struct EditTransactionSheetTests {
 
     @Test
     func canSubmit_nilAmount_returnsFalse() {
-        let result = EditTransactionSheet.isFormValid(
+        let result = EditTransactionLogic.isFormValid(
             name: "Groceries",
             amount: nil,
             isLoading: false
@@ -101,7 +64,7 @@ struct EditTransactionSheetTests {
 
     @Test
     func canSubmit_negativeAmount_returnsFalse() {
-        let result = EditTransactionSheet.isFormValid(
+        let result = EditTransactionLogic.isFormValid(
             name: "Groceries",
             amount: -10,
             isLoading: false
@@ -112,7 +75,7 @@ struct EditTransactionSheetTests {
 
     @Test
     func canSubmit_isLoading_returnsFalse() {
-        let result = EditTransactionSheet.isFormValid(
+        let result = EditTransactionLogic.isFormValid(
             name: "Groceries",
             amount: 50,
             isLoading: true
@@ -123,7 +86,7 @@ struct EditTransactionSheetTests {
 
     @Test
     func canSubmit_nameWithLeadingTrailingSpaces_returnsTrue() {
-        let result = EditTransactionSheet.isFormValid(
+        let result = EditTransactionLogic.isFormValid(
             name: "  Groceries  ",
             amount: 50,
             isLoading: false
@@ -164,40 +127,40 @@ struct EditTransactionSheetTests {
     func shouldShowAlternateCurrency_differentCurrencies_returnsTrue() {
         let tx = Self.makeTransaction(originalCurrency: .eur)
 
-        #expect(EditTransactionSheet.shouldShowAlternateCurrency(for: tx, userCurrency: .chf) == true)
+        #expect(EditTransactionLogic.shouldShowAlternateCurrency(for: tx, userCurrency: .chf) == true)
     }
 
     @Test("Case 2: same currency — picker hidden")
     func shouldShowAlternateCurrency_sameCurrency_returnsFalse() {
         let tx = Self.makeTransaction(originalCurrency: .eur)
 
-        #expect(EditTransactionSheet.shouldShowAlternateCurrency(for: tx, userCurrency: .eur) == false)
+        #expect(EditTransactionLogic.shouldShowAlternateCurrency(for: tx, userCurrency: .eur) == false)
     }
 
     @Test("Case 3: mono-currency — picker hidden")
     func shouldShowAlternateCurrency_nilOriginal_returnsFalse() {
         let tx = Self.makeTransaction(originalCurrency: nil)
 
-        #expect(EditTransactionSheet.shouldShowAlternateCurrency(for: tx, userCurrency: .chf) == false)
+        #expect(EditTransactionLogic.shouldShowAlternateCurrency(for: tx, userCurrency: .chf) == false)
     }
 
     @Test("Case 1: initialAmount uses originalAmount for alternate currency")
     func initialAmount_alternateCurrency_returnsOriginalAmount() {
         let tx = Self.makeTransaction(amount: 95, originalAmount: 100, originalCurrency: .eur)
 
-        #expect(EditTransactionSheet.initialAmount(for: tx, userCurrency: .chf) == 100)
+        #expect(EditTransactionLogic.initialAmount(for: tx, userCurrency: .chf) == 100)
     }
 
     @Test("Case 3: initialAmount uses transaction.amount for mono-currency")
     func initialAmount_monoCurrency_returnsLineAmount() {
         let tx = Self.makeTransaction(amount: 42)
 
-        #expect(EditTransactionSheet.initialAmount(for: tx, userCurrency: .chf) == 42)
+        #expect(EditTransactionLogic.initialAmount(for: tx, userCurrency: .chf) == 42)
     }
 
     @Test("Case 5: mono-currency submit omits all currency metadata")
     func buildUpdate_monoCurrency_omitsCurrencyFields() {
-        let update = EditTransactionSheet.buildUpdate(
+        let update = EditTransactionLogic.buildUpdate(
             name: "Groceries",
             amount: 50,
             kind: .expense,
@@ -222,7 +185,7 @@ struct EditTransactionSheetTests {
             exchangeRate: Decimal(0.95)
         )
 
-        let update = EditTransactionSheet.buildUpdate(
+        let update = EditTransactionLogic.buildUpdate(
             name: "EU import",
             amount: 100,
             kind: .expense,
@@ -241,8 +204,8 @@ struct EditTransactionSheetTests {
     func shouldShowAlternateCurrency_isPure() {
         let tx = Self.makeTransaction(originalCurrency: .eur)
 
-        let first = EditTransactionSheet.shouldShowAlternateCurrency(for: tx, userCurrency: .chf)
-        let second = EditTransactionSheet.shouldShowAlternateCurrency(for: tx, userCurrency: .chf)
+        let first = EditTransactionLogic.shouldShowAlternateCurrency(for: tx, userCurrency: .chf)
+        let second = EditTransactionLogic.shouldShowAlternateCurrency(for: tx, userCurrency: .chf)
 
         #expect(first == true)
         #expect(first == second)
